refactor: 서비스 관련 로직 커스텀 훅으로 추출
